corrected notifications time

+++ chrome/android/java/src/org/chromium/chrome/browser/toolbar/top/ToolbarTablet.java
@@ -447,7 +447,7 @@ public class ToolbarTablet extends ToolbarLayout
     public void OnGetCurrentBalanceReport(String[] report) {}
 
     @Override
-    public void OnNotificationAdded(String id, int type, int timestamp,
+    public void OnNotificationAdded(String id, int type, long timestamp,
             String[] args) {
         if (mBraveRewardsNativeWorker != null) {
             mBraveRewardsNativeWorker.GetAllNotifications();
@@ -476,7 +476,7 @@ public class ToolbarTablet extends ToolbarLayout
     }
 
     @Override
-    public void OnGetLatestNotification(String id, int type, int timestamp,
+    public void OnGetLatestNotification(String id, int type, long timestamp,
             String[] args) {}
 
     @Override
